Adds line space to EOF Renames cache and network packages to local and remote Updates RemoteDataSource classes - moves the mapping inside the runCatching lambda to catch malformed json responses
