+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/data/remote/company/CompanyApi.kt
@@ -1,12 +1,14 @@
-package com.seancoyle.feature.launch.implementation.data.network.company
+package com.seancoyle.feature.launch.implementation.data.remote.company
 
 import retrofit2.http.GET
 
 internal interface CompanyApi {
+
     companion object {
         const val COMPANY_URL = "/v4/company"
     }
 
     @GET(COMPANY_URL)
-    suspend fun getCompany(): CompanyDto?
+    suspend fun getCompany(): CompanyDto
+
 }
