+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/data/remote/launch/LaunchApi.kt
@@ -1,10 +1,11 @@
-package com.seancoyle.feature.launch.implementation.data.network.launch
+package com.seancoyle.feature.launch.implementation.data.remote.launch
 
 import com.seancoyle.feature.launch.implementation.domain.model.LaunchOptions
 import retrofit2.http.Body
 import retrofit2.http.POST
 
 internal interface LaunchApi {
+
     companion object {
         private const val URL = "/v4/launches/query"
     }
@@ -12,5 +13,6 @@ internal interface LaunchApi {
     @POST(URL)
     suspend fun getLaunches(
         @Body options: LaunchOptions
-    ): LaunchesDto?
+    ): LaunchesDto
+
 }
