+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/data/repository/company/CompanyRemoteDataSource.kt
@@ -5,5 +5,7 @@ import com.seancoyle.core.common.result.LaunchResult
 import com.seancoyle.feature.launch.api.domain.model.Company
 
 internal interface CompanyRemoteDataSource {
+
     suspend fun getCompanyApi(): LaunchResult<Company, DataError>
+
 }
