+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/data/repository/launch/LaunchPreferencesDataSource.kt
@@ -5,10 +5,13 @@ import com.seancoyle.feature.launch.api.domain.model.LaunchPrefs
 import com.seancoyle.feature.launch.api.domain.model.LaunchStatus
 
 interface LaunchPreferencesDataSource {
+
     suspend fun saveLaunchPreferences(
         order: Order,
         launchStatus: LaunchStatus,
         launchYear: String
     )
+
     suspend fun getLaunchPreferences(): LaunchPrefs
+
 }
