+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/data/repository/launch/LaunchRemoteDataSource.kt
@@ -6,5 +6,9 @@ import com.seancoyle.feature.launch.api.domain.model.LaunchTypes
 import com.seancoyle.feature.launch.implementation.domain.model.LaunchOptions
 
 internal interface LaunchRemoteDataSource {
-    suspend fun getLaunches(launchOptions: LaunchOptions): LaunchResult<List<LaunchTypes.Launch>, RemoteError>
+
+    suspend fun getLaunches(
+        launchOptions: LaunchOptions
+    ): LaunchResult<List<LaunchTypes.Launch>, RemoteError>
+
 }
