+++ feature/launch/implementation/src/main/java/com/seancoyle/feature/launch/implementation/di/launch/LaunchPreferencesDataStoreProtoModule.kt
@@ -6,7 +6,7 @@ import androidx.datastore.core.DataStoreFactory
 import androidx.datastore.dataStoreFile
 import com.seancoyle.core.common.di.ApplicationScope
 import com.seancoyle.core.datastore_proto.LaunchPreferencesProto
-import com.seancoyle.feature.launch.implementation.data.cache.launch.LaunchPreferencesSerializer
+import com.seancoyle.feature.launch.implementation.data.local.launch.LaunchPreferencesSerializer
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
