+++ feature/launch/implementation/src/test/java/com/seancoyle/feature/launch/implementation/data/network/company/CompanyRemoteDataSourceImplTest.kt
@@ -5,6 +5,8 @@ import com.seancoyle.core.common.result.DataError.RemoteError
 import com.seancoyle.core.common.result.LaunchResult
 import com.seancoyle.feature.launch.implementation.data.mapper.CompanyMapper
 import com.seancoyle.feature.launch.implementation.data.mapper.RemoteErrorMapper
+import com.seancoyle.feature.launch.implementation.data.remote.company.CompanyApi
+import com.seancoyle.feature.launch.implementation.data.remote.company.CompanyRemoteDataSourceImpl
 import com.seancoyle.feature.launch.implementation.data.repository.company.CompanyRemoteDataSource
 import com.seancoyle.feature.launch.implementation.util.TestData.companyDto
 import com.seancoyle.feature.launch.implementation.util.TestData.companyModel
