+++ feature/launch/implementation/src/test/java/com/seancoyle/feature/launch/implementation/util/TestData.kt
@@ -15,12 +15,12 @@ import com.seancoyle.feature.launch.api.domain.model.LaunchTypes
 import com.seancoyle.feature.launch.api.domain.model.Links
 import com.seancoyle.feature.launch.api.domain.model.Rocket
 import com.seancoyle.feature.launch.api.domain.model.RocketWithMission
-import com.seancoyle.feature.launch.implementation.data.network.company.CompanyDto
-import com.seancoyle.feature.launch.implementation.data.network.launch.LaunchDto
-import com.seancoyle.feature.launch.implementation.data.network.launch.LaunchesDto
-import com.seancoyle.feature.launch.implementation.data.network.launch.LinksDto
-import com.seancoyle.feature.launch.implementation.data.network.launch.PatchDto
-import com.seancoyle.feature.launch.implementation.data.network.launch.RocketDto
+import com.seancoyle.feature.launch.implementation.data.remote.company.CompanyDto
+import com.seancoyle.feature.launch.implementation.data.remote.launch.LaunchDto
+import com.seancoyle.feature.launch.implementation.data.remote.launch.LaunchesDto
+import com.seancoyle.feature.launch.implementation.data.remote.launch.LinksDto
+import com.seancoyle.feature.launch.implementation.data.remote.launch.PatchDto
+import com.seancoyle.feature.launch.implementation.data.remote.launch.RocketDto
 import com.seancoyle.feature.launch.implementation.domain.model.LaunchOptions
 import com.seancoyle.feature.launch.implementation.domain.model.Options
 import com.seancoyle.feature.launch.implementation.domain.model.Populate
